feat: add init command basic structure
also added the project structure as part of doing it. 'log' is the package for logging any information into the terminal, 'validation' ensures the conditions for the CLI to work correctly (but doesn't manage how they must be resolved), 'errors' formulates the structure / content of customized error messages according to the program's logic. Lastly, 'constants' holds the constants (mostly strings) that are required on different commands or packages of the app. I had to come up with this structure because the way this language works is going to kill me