fix: fix action not executing the combine sections script properly
so literally I forgot I changed the script to be in python, and it was being executed using still golang. I changed that as it had to and did other necesary changes as adding the -f flag to the add README.md command, among other things. Will anyone ever read these?